Add workaround for race condition
At the moment we rely on the 'webrtcup' Janus event for our 'established'
event.  Turns out, this can be problematic, because on outgoing calls that can
come before the setRemoteDescription callback kicks in, which is where we
post the 'accepted' event.  As a result, we get the state changes in the
incorrect order.

This is just a stopgap, a cleaner fix is needed.