Never call getUserMedia ourselves
This simplifies our API and pushes that responsibility to the user. Why the change?
That's a good question.  I observed that we had quite a bit of code duplication in
Sylk because we want to call getUserMedia there, so we might as well make it a thing.
Less is more!