Don't return ended local/remote videotracks for a call
Since we use getSenders/getReceivers unless it is not supported,
a receiver or sender is present for video when we start a call
with video. If the remote anwers with audio and no video, the
receiver is still there, but the video is ended. In this case it
needs to be not present in the mediaStream when we do
getRemoteStreams or getLocalStreams.